works properly

+++ drawable.cpp
@@ -12,5 +12,5 @@ void Drawable::drawBall(QPainter &painter, double x, double y, int w, int h, QCo
     static_cast<int>(mainColor.green() * 0.80),
     static_cast<int>(mainColor.blue() * 0.80)));
     painter.setBrush(mainColor);
-    painter.drawEllipse(x, y, w, h);
+    painter.drawEllipse(static_cast<int>(x), static_cast<int>(y), w, h);
 }
